requirements.txt
flask
groq
sentence-transformers
hnswlib
numpy
python-dotenv



